Separate REST request response models. Model module.
