+++ core/gsb-api/src/lib.rs
@@ -1,28 +1,26 @@
 mod api;
+mod model;
 mod service;
 mod services;
 
 use crate::service::{DropMessages, StartBuffering, StartRelaying};
 use actix::prelude::*;
 use actix::ActorFutureExt;
-use actix::{Actor, Addr, Handler, MailboxError, StreamHandler};
+use actix::{Actor, Addr, Handler, StreamHandler};
 use actix_http::ws::CloseCode;
 use actix_http::{
     ws::{CloseReason, ProtocolError},
-    StatusCode,
 };
-use actix_web::{HttpResponse, ResponseError};
 use actix_web_actors::ws::{self, WebsocketContext};
 use flexbuffers::{BuilderOptions, MapReader, Reader};
 use futures::FutureExt;
 use serde::{Deserialize, Serialize};
 use service::Service;
-use services::{BindError, FindError, Services, UnbindError};
-use thiserror::Error;
-use ya_client_model::ErrorMessage;
+use services::{Services};
 use ya_service_api_interfaces::Provider;
 
 pub const GSB_API_PATH: &str = "gsb-api/v1";
+
 pub struct GsbApiService;
 
 impl GsbApiService {
@@ -43,92 +41,6 @@ impl GsbApiService {
 }
 
 pub(crate) type GsbError = ya_service_bus::Error;
-
-#[derive(Error, Debug)]
-pub(crate) enum GsbApiError {
-    #[error("Bad request: {0}")]
-    BadRequest(String),
-    #[error("Not found: {0}")]
-    NotFound(String),
-    #[error("Internal error: {0}")]
-    InternalError(String),
-}
-
-impl From<BindError> for GsbApiError {
-    fn from(error: BindError) -> Self {
-        match error {
-            BindError::DuplicatedService(_) => Self::BadRequest(error.to_string()),
-            BindError::InvalidService(_) => Self::BadRequest(error.to_string()),
-        }
-    }
-}
-
-impl From<UnbindError> for GsbApiError {
-    fn from(error: UnbindError) -> Self {
-        match error {
-            UnbindError::ServiceNotFound(_) => Self::NotFound(error.to_string()),
-            UnbindError::InvalidService(_) => Self::BadRequest(error.to_string()),
-            UnbindError::UnbindFailed(_) => Self::InternalError(error.to_string()),
-        }
-    }
-}
-
-impl From<FindError> for GsbApiError {
-    fn from(error: FindError) -> Self {
-        match error {
-            FindError::EmptyAddress => Self::BadRequest(error.to_string()),
-            FindError::ServiceNotFound(_) => Self::NotFound(error.to_string()),
-        }
-    }
-}
-
-impl From<GsbError> for GsbApiError {
-    fn from(value: GsbError) -> Self {
-        GsbApiError::InternalError(format!("GSB error: {value}"))
-    }
-}
-
-impl From<MailboxError> for GsbApiError {
-    fn from(value: MailboxError) -> Self {
-        GsbApiError::InternalError(format!("Actix error: {value}"))
-    }
-}
-
-impl From<serde_json::Error> for GsbApiError {
-    fn from(value: serde_json::Error) -> Self {
-        GsbApiError::InternalError(format!("Serialization error {value}"))
-    }
-}
-
-impl From<actix_web::Error> for GsbApiError {
-    fn from(value: actix_web::Error) -> Self {
-        GsbApiError::InternalError(format!("Actix error: {value}"))
-    }
-}
-
-impl ResponseError for GsbApiError {
-    fn status_code(&self) -> StatusCode {
-        match *self {
-            Self::BadRequest(_) => StatusCode::BAD_REQUEST,
-            Self::NotFound(_) => StatusCode::NOT_FOUND,
-            Self::InternalError(_) => StatusCode::INTERNAL_SERVER_ERROR,
-        }
-    }
-
-    fn error_response(&self) -> actix_web::HttpResponse<actix_http::body::BoxBody> {
-        match self {
-            GsbApiError::BadRequest(message) => {
-                HttpResponse::BadRequest().json(ErrorMessage::new(message))
-            }
-            GsbApiError::NotFound(message) => {
-                HttpResponse::NotFound().json(ErrorMessage::new(message))
-            }
-            GsbApiError::InternalError(message) => {
-                HttpResponse::InternalServerError().json(ErrorMessage::new(message))
-            }
-        }
-    }
-}
 
 #[derive(Message, Serialize, Deserialize, Debug)]
 #[rtype(result = "Result<(), anyhow::Error>")]
